+++ jsf/src/main/java/geekbrains/persist/Product.java
@@ -26,7 +26,7 @@ public class Product implements Serializable {
     private String name;
 
     @Column
-    private Long price;
+    private int price;
 
     @ManyToOne
     @NotNull(message = "Поле не должно быть пустым")
@@ -35,7 +35,7 @@ public class Product implements Serializable {
     public Product() {
     }
 
-    public Product(Long id, String name, Long price, Category category) {
+    public Product(Long id, String name, int price, Category category) {
         this.id = id;
         this.name = name;
         this.price = price;
